libqmi-glib,device: remove message version checks
When using the QMI_DEVICE_OPEN_FLAGS_VERSION_INFO flag to open the
QmiDevice, we would be doing automatic checks to validate whether a
given message that we're attempting to use is supposedly supported by
the specific QMI service version reported by the device.

This logic is broken, because our information about when each message
was introduced is far from good, and we've seen lots of cases where
the info we thought was good ended up not being good (and we manually
reverted the specific message supported version to 1.0 in those
cases).

Just stop doing this message version validation, and assume all are
supported, and if they aren't it will be the module returning an
error.

(cherry picked from commit 6e2da987b887b47783091fe1df7e28b48962bfb9)