uim: common "UIM session" TLV was added in 1.22
The "Session" TLV was added in 1.22 as a replacement of the older
"Session Information", because the format of the "application
identifier" field changed, so a full new set of methods was required.

While doing that, the new "Session" TLV was flagged with version 1.22
correctly, but a follow up commit (d83fbdef16) ended up breaking this.

Reverting d83fbdef16 would be enough, but let's fix this by modifying
the per-message TLV since tags instead, which is probably clearer
than keeping the since tag in the common reference type.