libqmi-glib,nas: fix the format of the NITZ information TLV
This TLV includes both a long and short operator name string, which
are encoded either in GSM-7 or UCS-2 depending on the "Name encoding"
field.

Therefore, we cannot assume we can read these fields as standard
strings; instead we will read the fields as GArrays of guint8 values.

The old format is deprecated, and the new format is defined with a new
TLV name.