libqmi-glib,message: attempt to read as GSM-7 or UCS-2 the strings
Due to known firmware bugs, the operator name string provided in the
'Service Provider Name' TLV or the 'Current PLMN' TLV description
string is not always in ASCII-7 format, sometimes it comes in e.g.
GSM-7.

In order to avoid problems in the clients reading these strings, we'll
attempt to read the strings as GSM-7 or UCS-2 if the initial UTF-8
encoding validation fails.

Fixes https://gitlab.freedesktop.org/mobile-broadband/libqmi/issues/6