libqmi-glib,nas: fix the format of the Home Network 3GPP2 TLV
This TLV includes a description string which is encoded in either
ASCII-7, GSM-7 or UCS-2 depending on the "Description encoding"
field.

Therefore, we cannot assume we can read these fields as standard
strings; instead we will read the fields as GArrays of guint8 values.

The old format is deprecated, and the new format is defined with a new
TLV name.

(cherry picked from commit a420431c6512911ad5ff636ade0ed047b265bce3)