qrtr-control-socket: new explicit peek_node()
The old get_node() wasn't returning a full reference, and that would
have break the qrtr_node_for_id() implementation because the node
object would no longer exist once the control socket inside the GTask
was unref-ed.

So cleanup this logic and setup a peek_node() that is transfer-none
and a get_node() that is transfer-full.