libqmi-glib,utils: assume non-printable UTF-8 is not expected
When a string comes in GSM7 instead of UTF-8, it may also happen that
it ends up being valid UTF-8 anyway. But, there may be cases that the
UTF-8 validity accepts non-printable characters, and that is something
we currently don't expect reported from the modem, so just go on and
assume that even if a string is valid UTF-8, it must contain all
printable UTF-8 characters, otherwise we'll attempt GSM-7 and UCS-2
decoding instead.

Fixes https://gitlab.freedesktop.org/mobile-broadband/ModemManager/issues/191