libqmi-glib,message: fixed sized strings may be suffixed with 0xFF
If the contents of a fixed sized string are shorter than the field
size (e.g. a 2-digit MNC in a 3-digit string), a 0xFF filled suffix
will be included.

Since the 0xFF suffix makes the full string not UTF-8 valid, the
method that reads the fixed sized strings is updated to allow
returning all the valid UTF-8 characters found at the beginning of the
string, if any, instead of returning an error.

Fixes https://gitlab.freedesktop.org/mobile-broadband/libqmi/-/issues/33