nas, qmicli: Added 5G signal info TLV's
Added 5G signal info and 5G signal info extended TLV's, which contain
RSRP, RSRQ and SNR when the 5G signal is active, if the signal
is not active the value -32768 is returned.

This has been tested with a Quectel RM500Q modem connected to a 5G
Amari callbox in 5G NSA mode, the values returned by the TLV's where
verified against the AT+ENG="servingcell" command.

    > qmicli -d /dev/cdc-wdm0 -p --nas-get-signal-info

    [/dev/cdc-wdm0] Successfully got signal info

    LTE:
            RSSI: '-47 dBm'
            RSRQ: '-5 dB'
            RSRP: '-70 dBm'
            SNR: '30.0 dB'

    5G:
            RSRP: '-75 dBm'
            SNR: '18.5 dB'
            RSRQ: '-11 dB'

    >mmcli -m 0 --command=+qeng=\"servingcell\"
    response: '+QENG: "servingcell","NOCONN"
    +QENG: "LTE","FDD",1,01,1A2D001,1,1575,3,3,3,1,-70,-9,-46,25,-
    +QENG:"NR5G-NSA",1,01,500,-75,180,-11'