qmi-codegen: (skip) return value in bundle getters and setters
The return value (TRUE or FALSE) is not required in the introspection
support because the method already throws an exception on failure (in
other words, if there is no exception the return value would always be
TRUE), so explicitly (skip) it.

This commit ends up triggering an API break in the bindings generated
via GObject introspection, because the getter and setter calls would
now have one less return argument.

E.g. instead of:
  success, imei = output.get_imei()
We should now do:
    imei = output.get_imei()

There is no API break in libqmi-glib.