qmi-codegen: (skip) return value in result field parser
The return value (TRUE or FALSE) is not required in the introspection
support because the method already throws an exception on failure (in
other words, if there is no exception the return value would always be
TRUE), so explicitly (skip) it.

This commit ends up triggering an API break in the bindings generated
via GObject introspection, because the field parser would now have no
return arguments.

E.g. instead of:
  success = output.get_result()
We should now do:
    output.get_result()

There is no API break in libqmi-glib.