loc: Add LOC Get/Set Engine Lock
This seems to allow blocking the location engine. When position sessions
are locked, there are no position reports sent. I'm not exactly sure
for what this is useful, but it's helpful to have the Get Engine Lock
message to make sure that no-one else has locked the engine.

The definitions have been translated from:
  https://source.codeaurora.org/quic/la/platform/vendor/qcom-opensource/location/tree/loc_api/loc_api_v02/location_service_v02.h?h=LA.BR.1.2.9.1-02310-8x16.0
  https://source.codeaurora.org/quic/la/platform/vendor/qcom-opensource/location/tree/loc_api/loc_api_v02/location_service_v02.c?h=LA.BR.1.2.9.1-02310-8x16.0