qmi-device: add net link management support
This allows us to create and delete net interfaces for qmi devices
using netlink. If mux_id is set to QMI_DEVICE_MUX_ID_AUTOMATIC when
creating a net interface for `base_ifname`, we search for the lowest N
such that `base_ifname.N` is not an available interface, and assign
mux_id N+1. If another mux_id is given, then we try to create a net
interface using the provided mux_id.
This mux ID should be used in the WDS Bind Mux Data Port
message.