libqrtr-glib: rename 'control socket' to 'bus'
The library should name the objects based on what the object
represents, not based on how the object is implemented. E.g. a
QrtrNode is the representation of a node in the QRTR bus, and so,
following the same reasoning, a QrtrBus is the representation of the
QRTR bus itself.