libqmi-glib,net-port-manager: new subclass for the qmi_wwan based logic
This commit introduces a new QmiNetPortManager implementation based on
the add_mux/del_mux interface provided by the qmi_wwan driver.

Several kernel releases after the add_mux/del_mux interface was
implemented, the driver was updated to include a new mux_id sysfs
attribute in the instantiated link interface. In kernel versions that
have support for this mux_id interface the implementation of the
QmiNetPortManager is reliable and almost fully featured (except for
the suggested interface prefix with creating a new link).

In kernel versions before the mux_id attribute was introduced, the
implementation is reliable only if a long-running process is the one
taking care of the link management (e.g. ModemManager). The
implementation cannot be reliable with ephimeral programs using
QmiDevice objects for single operations (e.g. qmicli).