libqmi-glib,endpoint-qrtr: only add allocation TLV on success
If the allocation/release operation fails, we don't include the TLV,
as it may not even have the correct info.

E.g. the release can only fail if the incoming message cannot be
correctly parsed, and in that case, the read cid will be 0, so there's
no point in adding that as a TLV in the returned error message.