libqrtr-glib,bus: new APIs to get the full list of nodes
The internal management of the nodes in the bus object is changed from
a hash table to a plain double linked list, so that implementing the
peek_nodes() and get_nodes() in the API is easier. When there aren't
many nodes, as is the usual case, there won't be loss in performance
between the lookup in the list vs the lookup in the hash table.