libqmi-glib,device: allow passing mux id as input for link deletion
When using the qmi_wwan driver without the new qmap/mux_id sysfs
attribute in the link interface, it is not possible to delete the
interface because the mux id is unknown.

Update the delete_link() method to allow providing a valid mux_id, or
UNBOUND if we want the library to try to find the correct one when
needed.