libqmi-glib,net-port-manager-qmiwwan: custom del_all_links() method
When using qmi_wwan as backend, there is no guarantee that the link
interface has the new 'qmap/mux_id' property, so there may be no way
to know which mux-id is associated to a given link, and therefore
there wouldn't any way to delete that link, as the 'del_mux' operation
requires the mux-id of the link to delete.

As a fallback, we implement a custom del_all_links() method in this
class, that will just iterate over all possible mux-ids, deleting them
one by one, and ignoring the operation result. The method will succeed
if there are no links listed once the operation has finished.