build: rmnet support depends on availability of kernel symbols
The if_links.h header exists since back 2012, but the new
ifla_rmnet_flags struct was added only in March 2018 (Linux v4.17).

Setup a configure time check to see if the struct exists in the header
or not, and build rmnet support if so.

Fixes https://gitlab.freedesktop.org/mobile-broadband/libqmi/-/issues/55