libqmi-glib,net-port-manager-qmiwwan: simplify add/del operations
The original logic assumed that after the sysfs add_mux/del_mux write
operation, the link interface would require some time to show up in
sysfs, and so we would attempt to always run one by one these
operations, so that e.g. we could keep track of what mux_id was
requested and what link name was generated.

But this doesn't seem to be true; add_mux/del_mux seem to guarantee
that as soon as the write operation finishes successfully, the new
link is already available (for add_mux) or the given link was already
removed (for del_mux).

So just simplify the logic, and run the whole add/del logic in a sync
way without forcing an artificial 250ms timeout to complete each
operation.