libmbim-glib,device: new link add/delete operations
The MBIM devices support multiplexing as part of the protocol itself,
and in the cdc_mbim driver this feature has been supported from the
very first version of the driver.

In the cdc_mbim driver, MBIM IP sessions are mapped to VLAN IDs
matching the MBIM session IDs, and so we can use netlink operations to
request adding or removing links as appropriate.

This operation is completely equivalent to the link management
operations done in libqmi when using RMNET, so the new net port
manager internal object running the netlink operations is highly based
on the QmiNetPortManagerRmnet class defined in libqmi-glib.

Unlike in the libqmi case, there are no other supported net link
management backends, just one, so the MbimNetPortManager is not an
abstract class, it is instead the class implementing support for the
single cdc_mbim backend.