libqmi-glib,device: plug memleak when QmiDevice is opened multiple times
On a QmiDevice re-open, if we're opening with the version info check
flag, we would be re-querying the device for the list of supported
services. That is fine, but we must make sure the array of supported
services from the last open operation is cleared before storing the
new one.