qmicli,dpm: add --dpm-open-port and --dpm-close-port actions
The open port action allows 3 optional TLVs, which we create from a
single string passed by the user.

Each of this TLVs is an array of items, and so the user can pass
multiple groups of keys in order to add multiple items on the same
array. E.g. this should build a request with 2 items in the hw data
port list:

 $ sudo qmicli -d qrtr://0 -p \
               --dpm-open-port="hw-data-ep-type=embedded,hw-data-ep-iface-number=1,hw-data-rx-id=1,hw-data-tx-id=2,
	                        hw-data-ep-type=embedded,hw-data-ep-iface-number=2,hw-data-rx-id=3,hw-data-tx-id=4"

Each array has a given set of keys to create its elements; if not all
keys for a given array element are given, the operation will fail.
E.g. in this case the second item lacks the 'hw-data-tx-id' field:

 $ sudo qmicli -d qrtr://0 -p \
               --dpm-open-port="hw-data-ep-type=embedded,hw-data-ep-iface-number=1,hw-data-rx-id=1,hw-data-tx-id=2,
	                        hw-data-ep-type=embedded,hw-data-ep-iface-number=2,hw-data-rx-id=3"
 error: unfinished hw data port item