build: require libgudev 232
The 232 libgudev version is also the one provided by Ubuntu 18.04 LTS,
and so we can now say that this LTS release is the last one we support
in newer libqmi releases. The previous Ubuntu 16.04 LTS is already
out of the standard 5-year support.

Also, this new version gives us support to use g_autoptr() in all
libgudev types.