libmbim-glib,ms-basic-connect-extensions: rework 'LTE attach status'
The "LTE Attach Status" message was relying on a bogus
MbimLteAttachStatus struct that was created by the library and passed
as output of the parser methods to the user.

This struct is completely avoidable, just by updating the messages to
include the fields directly in the response/notification fields.

We rename the method to 'LTE attach info' following this approach, and
compatibility symbols are included in order to avoid breaking API/ABI.