libmbim-glib,ms-uicc-low-level-access: adding cid 'uicc-apdu'
MSFT MBIM_CID_MS_UICC_APDU is used to send a command APDU to
a specific logical channel on the UICC and receive the
response. Sending entity has to ensure that the logical
channel was previously opened with uicc-open channel.
Refer section 10.1.1 of ETSI TS 102 221 TS for more details
on apdu encoding/decoding.