libmbim-glib,ms-uicc-low-level-access: adding cid 'uicc-reset'
MSFT MBIM_CID_MS_UICC_RESET is used to reset the UICC or
to query passthrough state of the function.
If host/application specifies passthrough action, the function
resets the UICC and upon UICC power up, treats the UICC as if
it were in a passthrough mode. If the passthrough action is
disabled during reset operation then upon power up, treats
the UICC as a regular Telecom UICC and expects a Telecom
UICC file system to be present on the UICC