build: allow disabling rmnet support at configure time
There may be systems using a new enough kernel that would have rmnet
support as seen in the installed kernel headers, but in which
CONFIG_RMNET was not really enabled, and so there is no true rmnet
support during runtime.

We now allow disabling rmnet support completely at configure time, so
that all that logic is completely skipped if we know the kernel won't
support it.