build: don't depend on kernel headers for the rmnet support
The rmnet support is no longer "auto"; it won't depend on whether we
detect the rmnet support in the kernel headers. Instead, we'll use a
fallback header defining the rmnet netlink support in case there are
no kernel headers installed or if the kernel headers don't have the
rmnet related symbols.