libmbim-glib,ms-basic-connect-extensions: adding cid 'version'
The Version CID implementation.
For MBB drivers that support MBIM Microsoft extension 2.0 or above,
MBIM_CID_VERSION is a mandatory command for exchanging MBIM version
information between the host and the device. For in-market devices
with drivers that do not recognize this CID, the host will assume
and provide backward compatibility.