libmbim-glib,device: ensure CLOSED state always in mbim_device_close()
A mbim_device_close() should always finish with the MbimDevice in
closed state, regardless of whether there was an error reported or not
in the operation.

In other words, if any error is reported in mbim_device_close(), the
user can still assume the setup is fully closed even if some step
failed.