build,meson: Fix dependencies
In a recent commit (3b5d5862), `libqmi_glib_generated_dep` was
removed as public dependency from `libqmi_glib_dep`.

The reason behind this removal was the transitive dependency of
`libqmi_glib` over `libqmi_glib_generated`[0][1], necessary for some
targets, i.e `qmicli`, and the use of `libqmi_glib_dep` in
documentation generation[2]. `libqmi_glib_generated` is a static
library so the meson produced `gtkdoc` target crashed. This is
something that is fixed in meson[3] but not yet in 0.45.1, the
current meson version used in `libqmi`.

To fix this, `libqmi_glib_compat` builds the whole
`libqmi_glib_generated` in itself, which is also fully included
in `libqmi_glib`, so it isn't a transitive dependency anymore.
The old `libqmi_glib_generated_dep` has been replaced by a new
dependency that includes everything but the library itself, so
it can be freely included. This way, `gtkdoc` will not try to
resolve a static library.

[0] https://gitlab.freedesktop.org/mobile-broadband/libqmi/-/blob/311af7e04cd16a732ec799969b1e8d9c5334ac27/src/libqmi-glib/meson.build#L95
[1] https://gitlab.freedesktop.org/mobile-broadband/libqmi/-/blob/311af7e04cd16a732ec799969b1e8d9c5334ac27/src/libqmi-glib/meson.build#L127
[2] https://gitlab.freedesktop.org/mobile-broadband/libqmi/-/blob/d780ddf3b33791b072bfe4a8313054677f6d6900/docs/reference/libqmi-glib/meson.build#L73
[3] https://github.com/mesonbuild/meson/issues/3935

Refs: https://gitlab.freedesktop.org/mobile-broadband/libqmi/-/merge_requests/273#note_1016004