build,meson: Fix missing symbols
The whole `libqmi-glib-generated` is set to be linked when building
`libqmi-glib-compat`. Then the whole `libqmi-glib-compat` is also
set to be linked when building `libqmi-glib`. This is:

 `libqmi-glib-generated` -> `libqmi-glib-compat` -> `libqmi-glib`

However, the symbols of `libqmi-glib-generated` are not included in
`libqmi-glib`. This is because meson doesn't generate the proper
chain when building different static libraries.

This is issue is also open in meson[0] but it is fixed in recent
meson versions.

To also include the symbols from `libqmi-glib-generated` in the
current meson version used in `libqmi`, it is now set to be linked
in `libqmi-glib`.

[0] https://github.com/mesonbuild/meson/issues/5343