mbimcli,ms-basic-connect-extensions: switch 16 bit reader to bcd reader
The MBIM version numbers given in the MBIM Version command (e.g.
"1.12" encoded as 0x12 0x01) should be read as 2 separate
bytes (e.g. 0x01 and 0x12), not as 2 separate 16bit values. Therefore,
there is not much need on having a specific 16 bit reader.

Instead, given that the Version is BCD encoded, we can have a BCD
string reader instead, which is much more useful as it allows us to
skip needing the custom BCD value building from the integer read in
base 10.