libmbim-glib,device: new 'MS_MBIMEX_V2' open flag
The Microsoft MBIM extensions for the Basic Connect service define API
breaks in the MBIM messages. In order to properly use the correct API,
the first interaction between host and device should define which is
the version of the protocol extensions agreed to use between both.

When the MBIM session is opened, the process managing the modem in the
host will specify the maximum version it supports, and the modem will
reply with the maximum version that both support.

E.g.:
 * If the host does not run the initialization sequence, both assume
   v1.0 is the version to use (backwards compatibility).
 * If the host reports that it supports MBIMEx v2.0 but the modem
   does not support any version other than 1.0, it will return an
   error, and both assume v1.0 is the version to use.
 * If the host reports MBIMEx v2.0 and the modem also supports that
   version (may or may not support other newer ones as well), it will
   return MBIMEx v2.0 as supported and it will be the agreed version
   to use.
 * If the host reports MBIMEx v3.0, but the modem only supports v2.0,
   it will return MBIMEx v2.0 as supported and it will be the agreed version
   to use.

Based on a patch written by:
Som_SP <somashekhar.puttagangaiah@intel.com>.