libmbim-glib,device: new methods to manage the MBIMEx version
The exchange of the MBIMEx version to use between host and device is
exclusively done via the MBIM_DEVICE_OPEN_FLAGS_MS_MBIMEX_V2 flag in
the mbim_device_open_full() request. Once this request finishes, the
device instance will know which is the MBIMEx version agreed, and the
user can gather this same information with the new version getter
method.

In addition to the version getter, a simple version checker method is
also implemented, which allows users of the API to quickly check
whether the agreed version is equal or greater than the given one, and
this helps writing clear code doing the checks.

A new setter method is also provided for the device object, which is
supposed to be used when the user already knows which specific version
has been agreed between host and device. The setter doesn't do any
MBIM interaction with the real device, it just updates its internal
records of which the current MBIMEx version is.