mbim-codegen: report UNSUPPORTED error when printing unknown CID
If we receive a request to print the contents of a message with a CID
that is not supported in the library, return an UNSUPPORTED error.

This will also allow us to fallback print as v1 the Basic Connect
messages from v2 which we don't support yet, or those which don't have
a counterpart in the v2 set.