build-aux,codegen: define new 'ms-struct-array' type
The new 'ms-struct-array' type is a complex type that has one single
purpose for now: supporting the Signal State v2 message defined by
Microsoft.

This is NOT a generic array type.

The array is defined as 2 uint32 values in the fixed buffer, one
giving the offset to where the intermediate struct starts, and the
next one with the full size of the whole intermediate struct (which we
read but silently ignore).

The intermediate struct has a first uint32 value with the amount of
elements in the array, and then the array itself follows. The elements
of the array are structs with fixed size.

If the number of elements in the array is 0, the pointer to the
intermediate struct will be 0x00000000.

Microsoft, next time please use the standard way of defininig struct
arrays in MBIM, i.e. by the means of using arrays of OL pairs (our
'ref-struct-array' type in libmbim).