ms-basic-connect-extensions: implement wake-reason
This CID is used by host to query the modem for its wake reason when
power up because of device wake. If the device supports this function,
but fails the query from host, it means the power up is not caused by
a wake from the device.

Includes updates by Aleksander Morgado to fix coding style issues,
message definitions add missing documentation items.