build-aux,codegen: generate support for all versions of same service together
There are service updates that will re-use parts of the previous
versions of the same service (e.g. struct types and such). In order to
provide an easy way to reference all those common types among versions
of the same service, we'll just build all of them together in the same
source files.