ms-basic-connect-v3: fix MbimTai struct definition
The PLMN field is not a single 32 bit value, it's really 2 separate
guint16 values. This is important because the endianness change in BE
systems would be different.

Also, the format of the MCC and MNC inside the PLMN fields is very
specific to the implementation (BCD encoded, MNC PCS digit indicated
in most significant bit of the MNC field...).