libmbim-glib: new MbimTlv generic type
The Microsoft Extensions define a new generic TLV type that allows
including any kind of data in it: from strings, to arrays of uints, to
more complex types.

We define support for the generic 'tlv' type and the generic
'tlv-list' type, which allow implementing the "unnamed IEs" that are
appended to some new messages in the MS extensions.

And we also define support for a more specific 'tlv-string', used
exclusively when the TLVs are part of the named fields, so that we can
provide strings in our libmbim-glib APIs, instead of requiring the
user to build MbimTlv instances.