build: Fix double quotes in rules file
`QMI_USERNAME` is quoted twice in the
`76-qmi-proxy-device-ownership.rules` rules file. This is because
the definition is used in two files.

The first one is `config.h` file, that is used in source code files
and needs to be explicitly quoted. The second one is the rules file,
mentioned above that already has quotes in the template file.

The explicit quotes have been removed in the rules file generation.