libmbim-glib,compat: deprecate the MbimLteAttachStatus type
We introduce a new MbimLteAttachStatusDeprecated symbol that isn't
really deprecated, which we use in our own APIs, and we typedef
MbimLteAttachStatus to it.

The only method in which we couldn't use MbimLteAttachStatusDeprecated
was the autoptr cleanup setup, so we just explicitly request to ignore
deprecation warnings on that one in both gcc and clang.