build-aux,codegen: allow defining per-container "compat" contexts
The compat methods need to be able to return variables in
transfer-none, i.e. still owned by the container itself. In order to
do so in compat methods, we will create a custom compat context that
is owned by the container, so that when the container is disposed, so
is the compat context.