compat: fix 'GERAN Info' TLV return in 'NAS Get Cell Location Info'
The 'GERAN Info' TLV returned a "cell" GArray with array elements of
the wrong format. This was fixed in the 'GERAN Info v2' update, but
the compat method still needs to be able to return a GArray

We could return an empty GArray and be done with it, but providing a
compat array with the old element type isn't a big deal either, so
we try to do that.

This avoids breaking users that were correctly assuming the output
GArray was always set if the TLV getter didn't fail.