qmi-firmware-update: always validate type of TTY before using it
When listing TTYs or waiting for TTYs, always explicitly specify the
exact mode the modem should be in, so that we don't get the wrong
device.

E.g. after a reset request, always wait for a TTY with the device in
download mode.

The modem mode is automatically detected based on the amount of USB
interfaces the device exposes. If the device has a single USB
interface, then it's in download mode, otherwise it's in modem mode.
A vid:pid match could also be possible, but it would be harder to
maintain. Let's use this approach for now and see how it goes.