libqmi-glib,net-port-manager: use unaligned netlink attribute length
The attribute length should not include padding.

The kernel will validate the length when it can, and will complain
if we include padding of 8bit or 16bit values.  In theory, there is
also a possibility that the aligned length exceeds an attribute
specific limit set by kernel policy, while the actual value is
within limits.

Fixes: 4085a2d6b771 ("qmi-device: add net link management support")
Signed-off-by: Bjørn Mork <bjorn@mork.no>