intel-firmware-update: implement ModemReboot version 2
Existing intel firware update doesn't support parameters for customized
reboot for firmware flashing even though the original intention of the
command is for firmware update. This extension is required to add
parameters to control boot mode when the modem firmware is rebooted
for flashing.

This extension will be available only if MBIMEx version is 2 or above.

Updated by Aleksander Morgado according to reviews in
https://gitlab.freedesktop.org/mobile-broadband/libmbim/-/merge_requests/134