build-aux,codegen: print the '###' hidden field contents in a single place
Instead of checking whether personal info should be shown or not for
each field type in the message or struct processing, do it just once
at the very beginning, and only if there are fields that may be hidden
in the message.

Then, if we should not show the content of the field, we right away
print '###', and if we should show the content of the field, we reuse
the same logic we had before.