libqmi,qmicli: support Foxconn FCC authentication v2 format
This new message reuses the same 0x5571 DMS command as the original
one, but instead of one single magic number, the command expects first
a magic string as TLV 0x01 and then a magic number as TLV 0x02.

The prerequisite added to TLV 0x02 is so that it's ordered last (order
of the TLVs matters in this operation) nothing else.

Based on a MBIM traffic dump retrieved when using the lenovo-wwan-dpr
tool by Thilo-Alexander Ginkel <thilo@ginkel.com>.